extract DataDog setup into dedicated class
